requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
whitenoise==6.1.0
django-cleanup==6.0.0
django-livesync==0.5
djangorestframework==3.13.1
mssql-django==1.1.3
imutils==0.5.4
numpy
opencv-python-headless
Pillow==9.1.1
pipreqs==0.4.11
psycopg2==2.9.3
pyodbc==4.0.32
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9